requirements.txt
openai==1.15.0
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0
openai-agents==0.2.0
nest-asyncio==1.5.8